versioning change for jdk
git-svn-id: https://svn.apache.org/repos/asf/tomcat/tc5.5.x/trunk@992407 13f79535-47bb-0310-9956-ffa450edef68

+++ container/webapps/admin/WEB-INF/classes/org/apache/webapp/admin/connector/SaveConnectorAction.java
@@ -490,7 +490,7 @@ public final class SaveConnectorAction extends Action {
                     } catch (NumberFormatException e) {
                         mBServer.setAttribute(coname,
                                 new Attribute("SSLVerifyDepthText",
-                                        Integer.valueOf(10)));
+                                        Integer.valueOf(10).intValue()));
                     }
                 
                 String sSLCACertificateFile = cform.getSSLCACertificateFile();
